+++ packages/@orchidui-vue/src/Form/Input/OcInput.js
@@ -1,3 +1,4 @@
 import Input from "./OcInput.vue";
+import InputOption from "./OcInputOption.vue";
 
-export { Input };
+export { Input, InputOption };
