+++ packages/@orchidui-vue/src/Form/Select/OcSelect.js
@@ -1,5 +1,5 @@
 import Select from "./OcSelect.vue";
 import Option from "./OcOption.vue";
 import GroupOptions from "./OcGroupOptions.vue";
-import InputSelect from "./OcInputSelect.vue";
-export { Select, Option, GroupOptions, InputSelect };
+
+export { Select, Option, GroupOptions };
